removed unused comments, replaced store and reducer with RTK files, updated file paths, removed unused testing files
