+++ src/app/__tests__/ButtonContainer.test.tsx
@@ -4,10 +4,8 @@ import '@testing-library/jest-dom/extend-expect';
 import { TextEncoder } from 'util';
 global.TextEncoder = TextEncoder;
 import ButtonsContainer from '../containers/ButtonsContainer';
-// import { useStoreContext } from '../store';
 import userEvent from '@testing-library/user-event';
-// import { toggleMode } from '../actions/actions'; 
-import { toggleMode, mainSlice } from '../RTKslices';
+import { toggleMode, mainSlice } from '../slices/mainSlice';
 import { useDispatch, Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
