+++ src/app/components/Action.tsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 import ReactHover, { Trigger, Hover } from 'react-hover';
-import { changeView, changeSlider } from '../RTKslices';
+import { changeView, changeSlider } from '../slices/mainSlice';
 import { ActionProps, OptionsCursorTrueWithMargin } from '../FrontendTypes';
 import { useDispatch } from 'react-redux';
