+++ src/app/components/Tutorial.tsx
@@ -9,7 +9,7 @@ import { TutorialProps, TutorialState, StepsObj } from '../FrontendTypes';
 import { Button } from '@mui/material';
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
 const { Steps } = require('intro.js-react'); //Must be required in. This enables compatibility with TS. If imported in, throws ts error of not rendering steps as a class component correctly. The package 'intro.js-react' is small React wrapper around Intro.js. The wrapper provides support for both steps and hints. https://introjs.com/docs/
-import { setCurrentTabInApp, tutorialSaveSeriesToggle } from '../RTKslices'
+import { setCurrentTabInApp, tutorialSaveSeriesToggle } from '../slices/mainSlice'
 import { useDispatch, useSelector } from 'react-redux';
 /*
   This is the tutorial displayed when the "How to use" button is clicked
