--- src/app/constants/actionTypes.ts
@@ -1,25 +1,0 @@
-// export const TOGGLE_MODE = 'TOGGLE_MODE';
-// export const MOVE_BACKWARD = 'MOVE_BACKWARD';
-// export const MOVE_FORWARD = 'MOVEFORWARD';
-// export const IMPORT = 'IMPORT';
-// export const EMPTY = 'EMPTY';
-// export const CHANGE_VIEW = 'CHANGE_VIEW';
-// export const CHANGE_SLIDER = 'CHANGE_SLIDER';
-// export const SET_PORT = 'SET_PORT';
-// export const PAUSE = 'PAUSE';
-// export const PLAY = 'PLAY';
-// export const INITIAL_CONNECT = 'INITIAL_CONNECT';
-// export const NEW_SNAPSHOTS = 'NEW_SNAPSHOTS';
-// export const SET_TAB = 'SET_TAB';
-// export const DELETE_TAB = 'DELETE_TAB';
-// export const NO_DEV = 'NO_DEV';
-// export const TOGGLE_EXPANDED = 'TOGGLE_EXPANDED';
-// export const LAUNCH_CONTENT = 'LAUNCH_CONTENT';
-// export const SLIDER_ZERO = 'SLIDER_ZERO';
-// export const ON_HOVER = 'ON_HOVER';
-// export const ON_HOVER_EXIT = 'ON_HOVER_EXIT';
-// export const SAVE = 'SAVE';
-// export const DELETE_SERIES = 'DELETE_SERIES';
-// export const SET_CURRENT_LOCATION = 'SET_CURRENT_LOCATION';
-// export const SET_CURRENT_TAB_IN_APP = 'SET_CURRENT_TAB_IN_APP';
-// export const TUTORIAL_SAVE_SERIES_TOGGLE = 'TUTORIAL_SAVE_SERIES_TOGGLE';
